+++ js/data.js
@@ -1,0 +1,17 @@
+const COMMENTS = ['В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.', 'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+const NAMES = ['Иван', 'Пётр', 'Анастасия', 'Николай', 'Мария', 'Егор', 'Александр', 'Екатерина'];
+const MAX_LIKES = 200;
+const MIN_LIKES = 15;
+const MAX_COMMENT = 30;
+const MAX_PHOTO = 25;
+const ID = [];
+export {
+  COMMENTS,
+  NAMES,
+  MAX_LIKES,
+  MIN_LIKES,
+  MAX_COMMENT,
+  MAX_PHOTO,
+  ID
+};
